requirements.txt
numpy
matplotlib
scipy
cupy-cuda12x
control
opencv-python
sqlite
numba
scikit-learn
